Day 12 Complete
Super easy day. It looked a bit harder than usual at first, but it was just a recursive function that traverses a node tree.  Problem 2 was the same as problem 1 just with a single modification that took two lines of code, so I only have a problem 2 output on this one.